Migrate testsuite/testcases/src/test/cases/spoon-feed-reader
- There is no need to use golden input/output. We can assert the yylex() of the scanner, with the input as string.
- This allows to remove the main() method in the scanners, we just need to create the good data in the test, which is 2 lines of code.
