--- testsuite/testcases/src/test/cases/spoon-feed-reader/.gitignore
@@ -1,1 +1,0 @@
-SpoonFeedReader*.java
