--- testsuite/testcases/src/test/cases/spoon-feed-reader/SpoonFeedReader-0.output
@@ -1,1 +1,0 @@
-12345678𐌀
