--- testsuite/testcases/src/test/cases/spoon-feed-reader/SpoonFeedReaderF-0.input
@@ -1,1 +1,0 @@
-12345678𐌀
